Revert "spi: altera: fix size mismatch on 64 bit processors"
This reverts commit 1f8703013599ae37ce6676146fc1ab774f51737d.

+++ drivers/spi/spi-altera.c
@@ -49,7 +49,7 @@ struct altera_spi {
 	int len;
 	int count;
 	int bytes_per_word;
-	u32 imr;
+	unsigned long imr;
 
 	/* data buffers */
 	const unsigned char *tx;
